Keep speaker identity regressions lint-clean
